Adds test data (#7, #14)
