+++ src/main/resources/application-test.properties
@@ -1,7 +1,7 @@
 # ====================================
 #       Data Source Config
 # ====================================
-spring.datasource.url=jdbc:mysql://127.0.0.1:3306/funbank_test?createDatabaseIfNotExist=true
+spring.datasource.url=jdbc:mysql://127.0.0.1:33306/funbank_test?createDatabaseIfNotExist=true
 spring.datasource.username=root
 spring.datasource.password=123456
 spring.datasource.driver-class-name=com.mysql.cj.jdbc.Driver
